// src/app/characters.interface.ts
export interface ApiResponse{
    info: Information;
    results: Characters[];

}

export interface Information{
    count: number;
    pages: number;
    next: string;
    prev: string;
}

export interface Characters {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: Origin;
    location: Location;
    image: string;
    episode: string[];
    url: string;
    created: string;
}
  
export interface Origin {
    name: string;
    url: string;
}
  
export interface Location {
    name: string;
    url: string;
}